Working endpoints: -GET /api/code/N, - GET /code/N, - POST /api/code/new, - GET /code/new

+++ src/main/java/com/codeSharingPlatform/ApiController.java
@@ -16,7 +16,7 @@ public class ApiController {
     }
 
     @GetMapping("/code/{id}")
-    public Code getJsonObject(@PathVariable Integer id, HttpServletResponse response) {
+    public Code getJsonObject(@PathVariable Long id, HttpServletResponse response) {
         response.setHeader("Content-Type", "application/json");
         return repository.getCodeByID(id);
     }
