+++ src/main/java/com/codeSharingPlatform/CodeRepository.java
@@ -7,22 +7,22 @@ import java.util.LinkedHashMap;
 
 @Repository
 public class CodeRepository implements ObjectRepository<Code> {
-   private LinkedHashMap<Integer, Code> repository;
+   private LinkedHashMap<Long, Code> repository;
 
    public CodeRepository() {
        this.repository = new LinkedHashMap<>();
    }
 
    @Override
-    public Integer storeCode (Code code) {
-       Integer id = repository.size() + 1;
+    public Long storeCode (Code code) {
+       Long id = (long)repository.size() + 1;
        Code codeToPut = new Code(code.getCode(), LocalDateTime.now());
        repository.put(id, codeToPut);
        return id;
    }
 
    @Override
-    public Code getCodeByID(Integer id) {
+    public Code getCodeByID(Long id) {
        return repository.get(id);
    }
 }
