+++ src/main/java/com/codeSharingPlatform/ObjectRepository.java
@@ -3,6 +3,6 @@ package com.codeSharingPlatform;
 import java.util.LinkedHashMap;
 
 public interface ObjectRepository<T> {
-    Integer storeCode(T t);
-    T getCodeByID(Integer id);
+    Long storeCode(T t);
+    T getCodeByID(Long id);
 }
